requirements.txt
qrcode[pil]>=7.4
Pillow>=10.0

